+++ src/reference.hpp
@@ -10,11 +10,11 @@ class Reference {
 public:
     static void initialize_labels(std::istream & labels);
 
-    static void normalize_labels(double loss_normalizer);
+    static void normalize_labels(float loss_normalizer);
 
     //labels for each row of the dataset, according to the reference model
     //Will likely include misclassifications when compared to the true labels
-    static std::vector<double> labels;
+    static std::vector<float> labels;
 };
 
 #endif
